+++ addons/mail/models/mail_thread.py
@@ -455,39 +455,6 @@ class MailThread(models.AbstractModel):
     def _get_from_context_or_raise(self, thread_id):
         return self.browse(thread_id).exists()
 
-    # ------------------------------------------------------------
-    # FIELDS HELPERS
-    # ------------------------------------------------------------
-
-    @api.model
-    def _mail_get_partner_fields(self):
-        """ This method returns the fields to use to find the contact to link
-        when sending emails or notifications. Having partner is not always
-        necessary but gives more flexibility to notifications management. """
-        return [fname for fname in ('partner_id', 'partner_ids')
-                if fname in self]
-
-    def _mail_get_partners(self):
-        """ Give the default partners associated to customers.
-
-        :return dict: for each record ID, a res.partner recordsets being default
-          customers to contact;
-        """
-        partner_fields = self._mail_get_partner_fields()
-        return dict(
-            (record.id, self.env['res.partner'].union(*[record[fname] for fname in partner_fields]))
-            for record in self
-        )
-
-    @api.model
-    def _mail_get_primary_email_field(self):
-        """ Check if the "_primary_email" model attribute is correctly set and
-        matches an existing field, and return it. Otherwise return None. """
-        primary_email = getattr(self, '_primary_email', None)
-        if primary_email and primary_email in self._fields:
-            return primary_email
-        return None
-
     # ------------------------------------------------------
     # TRACKING / LOG
     # ------------------------------------------------------
