+++ addons/phone_validation/models/__init__.py
@@ -3,5 +3,6 @@
 
 from . import phone_blacklist
 from . import mail_thread_phone
+from . import models
 from . import res_partner
 from . import res_users
